+++ driver/main.cpp
@@ -923,6 +923,12 @@ void registerPredefinedTargetVersions() {
     break;
   case llvm::Triple::Emscripten:
     VersionCondition::addPredefinedGlobalIdent("Emscripten");
+    // Emscripten uses musl and libc++, so mimic a musl Linux platform:
+    VersionCondition::addPredefinedGlobalIdent("linux");
+    VersionCondition::addPredefinedGlobalIdent("Posix");
+    VersionCondition::addPredefinedGlobalIdent("CRuntime_Musl");
+    VersionCondition::addPredefinedGlobalIdent("CppRuntime_LLVM");
+    VersionCondition::addPredefinedGlobalIdent("CppRuntime_Clang"); // legacy
     break;
   default:
     if (triple.getEnvironment() == llvm::Triple::Android) {